Oracular GADDAG implementation. Still buggy
I have saved massive amounts of RAW by compressing the GADDAG via getting rid of all the wasted space in the child vector inside the nodes by storing those indexes on an int vector.

The implentation of future heautists was went the wrong way since even though I saved more than 30 seconds of processing time ( 4 m 34 s to 3 min 59 s)  when simulating 1000 matches, it made the average to drop by 9. And the reason is pinpointed. And will be fixed.